+++ src/main/java/com/nexters/keyme/auth/client/KakaoClient.java
@@ -1,6 +1,6 @@
 package com.nexters.keyme.auth.client;
 
-import com.nexters.keyme.member.dto.response.KakaoUserInfoResponse;
+import com.nexters.keyme.auth.dto.response.KakaoUserInfoResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
