+++ src/main/java/com/nexters/keyme/auth/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.nexters.keyme.auth.controller;
 
 
 import com.nexters.keyme.auth.service.AuthService;
-import com.nexters.keyme.member.dto.request.LoginRequest;
+import com.nexters.keyme.auth.dto.request.LoginRequest;
 import com.nexters.keyme.member.dto.response.MemberResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
