+++ src/main/java/com/nexters/keyme/auth/dto/response/AppleAuthKeysResponse.java
@@ -1,4 +1,4 @@
-package com.nexters.keyme.member.dto.response;
+package com.nexters.keyme.auth.dto.response;
 
 import lombok.Getter;
