+++ src/main/java/com/nexters/keyme/auth/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.nexters.keyme.auth.service;
 
 
-import com.nexters.keyme.member.dto.request.LoginRequest;
+import com.nexters.keyme.auth.dto.request.LoginRequest;
 import com.nexters.keyme.member.dto.response.MemberResponse;
 
 public interface AuthService {
