+++ src/main/java/com/nexters/keyme/common/config/WebConfig.java
@@ -1,12 +1,26 @@
 package com.nexters.keyme.common.config;
 
+import com.nexters.keyme.auth.resolver.UserInfoArgumentResolver;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import java.util.List;
+
 @Configuration
+@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
+
+    private final UserInfoArgumentResolver userInfoArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(userInfoArgumentResolver);
+        WebMvcConfigurer.super.addArgumentResolvers(resolvers);
+    }
 
     @Override
     public void addResourceHandlers(ResourceHandlerRegistry registry) {
