--- src/main/java/com/nexters/keyme/common/config/WebMvcConfig.java
@@ -1,22 +1,0 @@
-package com.nexters.keyme.common.config;
-
-import com.nexters.keyme.auth.resolver.UserInfoArgumentResolver;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebMvcConfig implements WebMvcConfigurer {
-    private final UserInfoArgumentResolver userInfoArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(userInfoArgumentResolver);
-        WebMvcConfigurer.super.addArgumentResolvers(resolvers);
-    }
-}
-
