+++ src/main/java/com/nexters/keyme/member/dto/response/MemberResponse.java
@@ -1,5 +1,6 @@
 package com.nexters.keyme.member.dto.response;
 
+import com.nexters.keyme.auth.dto.response.TokenResponse;
 import com.nexters.keyme.model.member.MemberEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
